Original commit by Lars:
Performance update

The ExternalData class now buffers the data it is pointing to in a deep copy.
This happens for all looped steps before they are triggered. This leads to deterministic output when running onTrigger on looped steps concurrently.
The new preTrigger() method invokes the update that deep copies data into the buffer.
This improves performance as steps do not have to lock their input data that was previously only stored externally in the output slot of the source.
Now that they only access the copy, there is no locking complications between steps that get input from the same source.